POC: RouteConfiguration and Routes DTO
