+++ Neos.Flow/Classes/Mvc/Routing/RouteConfiguration.php
@@ -1,0 +1,120 @@
+<?php
+namespace Neos\Flow\Mvc\Routing;
+
+/*
+ * This file is part of the Neos.Flow package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
+
+use Neos\Flow\Annotations as Flow;
+use Neos\Flow\Configuration\ConfigurationManager;
+use Neos\Flow\Mvc\Exception\InvalidRouteSetupException;
+
+/**
+ * The default router configuration
+ *
+ * @Flow\Scope("singleton")
+ * @api
+ */
+class RouteConfiguration
+{
+    /**
+     * @Flow\Inject
+     * @var ConfigurationManager
+     */
+    protected $configurationManager;
+
+    /**
+     * Array of routes to match against
+     *
+     * @var Routes
+     */
+    protected $routes;
+
+    /**
+     * true if route object have been created, otherwise false
+     *
+     * @var boolean
+     */
+    protected $routesCreated = false;
+
+    /**
+     * Returns a list of configured routes
+     *
+     * @return Routes
+     */
+    public function getRoutes()
+    {
+        $this->createRoutesFromConfiguration();
+        return $this->routes;
+    }
+
+    /**
+     * Creates \Neos\Flow\Mvc\Routing\Route objects from the injected routes
+     * configuration.
+     *
+     * @return void
+     * @throws InvalidRouteSetupException
+     */
+    protected function createRoutesFromConfiguration()
+    {
+        if ($this->routesCreated === true) {
+            return;
+        }
+        $this->initializeRoutesConfiguration();
+        $routes = [];
+        $routesWithHttpMethodConstraints = [];
+        foreach ($this->routesConfiguration as $routeConfiguration) {
+            $route = new Route();
+            if (isset($routeConfiguration['name'])) {
+                $route->setName($routeConfiguration['name']);
+            }
+            $uriPattern = $routeConfiguration['uriPattern'];
+            $route->setUriPattern($uriPattern);
+            if (isset($routeConfiguration['defaults'])) {
+                $route->setDefaults($routeConfiguration['defaults']);
+            }
+            if (isset($routeConfiguration['routeParts'])) {
+                $route->setRoutePartsConfiguration($routeConfiguration['routeParts']);
+            }
+            if (isset($routeConfiguration['toLowerCase'])) {
+                $route->setLowerCase($routeConfiguration['toLowerCase']);
+            }
+            if (isset($routeConfiguration['appendExceedingArguments'])) {
+                $route->setAppendExceedingArguments($routeConfiguration['appendExceedingArguments']);
+            }
+            if (isset($routeConfiguration['httpMethods'])) {
+                if (isset($routesWithHttpMethodConstraints[$uriPattern]) && $routesWithHttpMethodConstraints[$uriPattern] === false) {
+                    throw new InvalidRouteSetupException(sprintf('There are multiple routes with the uriPattern "%s" and "httpMethods" option set. Please specify accepted HTTP methods for all of these, or adjust the uriPattern', $uriPattern), 1365678427);
+                }
+                $routesWithHttpMethodConstraints[$uriPattern] = true;
+                $route->setHttpMethods($routeConfiguration['httpMethods']);
+            } else {
+                if (isset($routesWithHttpMethodConstraints[$uriPattern]) && $routesWithHttpMethodConstraints[$uriPattern] === true) {
+                    throw new InvalidRouteSetupException(sprintf('There are multiple routes with the uriPattern "%s" and "httpMethods" option set. Please specify accepted HTTP methods for all of these, or adjust the uriPattern', $uriPattern), 1365678432);
+                }
+                $routesWithHttpMethodConstraints[$uriPattern] = false;
+            }
+            $routes = $route;
+        }
+        $this->routes = Routes::fromArray($routes);
+        $this->routesCreated = true;
+    }
+
+    /**
+     * Checks if a routes configuration was set and otherwise loads the configuration from the configuration manager.
+     *
+     * @return void
+     */
+    protected function initializeRoutesConfiguration()
+    {
+        if ($this->routesConfiguration === null) {
+            $this->routesConfiguration = $this->configurationManager->getConfiguration(ConfigurationManager::CONFIGURATION_TYPE_ROUTES);
+        }
+    }
+}
