+++ Neos.Flow/Classes/Mvc/Routing/Router.php
@@ -12,7 +12,6 @@ namespace Neos\Flow\Mvc\Routing;
  */
 
 use Neos\Flow\Annotations as Flow;
-use Neos\Flow\Configuration\ConfigurationManager;
 use Neos\Flow\Http\Helper\RequestInformationHelper;
 use Neos\Flow\Http\Helper\UriHelper;
 use Neos\Flow\Log\Utility\LogEnvironment;
@@ -39,22 +38,9 @@ class Router implements RouterInterface
 
     /**
      * @Flow\Inject
-     * @var ConfigurationManager
-     */
-    protected $configurationManager;
-
-    /**
-     * @Flow\Inject
      * @var RouterCachingService
      */
     protected $routerCachingService;
-
-    /**
-     * Array containing the configuration for all routes
-     *
-     * @var array
-     */
-    protected $routesConfiguration = null;
 
     /**
      * Array of routes to match against
@@ -62,13 +48,6 @@ class Router implements RouterInterface
      * @var array
      */
     protected $routes = [];
-
-    /**
-     * true if route object have been created, otherwise false
-     *
-     * @var boolean
-     */
-    protected $routesCreated = false;
 
     /**
      * @var Route
@@ -89,18 +68,6 @@ class Router implements RouterInterface
     public function injectLogger(LoggerInterface $logger)
     {
         $this->logger = $logger;
-    }
-
-    /**
-     * Sets the routes configuration.
-     *
-     * @param array $routesConfiguration The routes configuration or NULL if it should be fetched from configuration
-     * @return void
-     */
-    public function setRoutesConfiguration(array $routesConfiguration = null)
-    {
-        $this->routesConfiguration = $routesConfiguration;
-        $this->routesCreated = false;
     }
 
     /**
@@ -151,29 +118,6 @@ class Router implements RouterInterface
     }
 
     /**
-     * Returns a list of configured routes
-     *
-     * @return array
-     */
-    public function getRoutes()
-    {
-        $this->createRoutesFromConfiguration();
-        return $this->routes;
-    }
-
-    /**
-     * Manually adds a route to the beginning of the configured routes
-     *
-     * @param Route $route
-     * @return void
-     */
-    public function addRoute(Route $route)
-    {
-        $this->createRoutesFromConfiguration();
-        array_unshift($this->routes, $route);
-    }
-
-    /**
      * Builds the corresponding uri (excluding protocol and host) by iterating
      * through all configured routes and calling their respective resolves()
      * method. If no matching route is found, an empty string is returned.
@@ -216,68 +160,5 @@ class Router implements RouterInterface
     public function getLastResolvedRoute()
     {
         return $this->lastResolvedRoute;
-    }
-
-    /**
-     * Creates \Neos\Flow\Mvc\Routing\Route objects from the injected routes
-     * configuration.
-     *
-     * @return void
-     * @throws InvalidRouteSetupException
-     */
-    protected function createRoutesFromConfiguration()
-    {
-        if ($this->routesCreated === true) {
-            return;
-        }
-        $this->initializeRoutesConfiguration();
-        $this->routes = [];
-        $routesWithHttpMethodConstraints = [];
-        foreach ($this->routesConfiguration as $routeConfiguration) {
-            $route = new Route();
-            if (isset($routeConfiguration['name'])) {
-                $route->setName($routeConfiguration['name']);
-            }
-            $uriPattern = $routeConfiguration['uriPattern'];
-            $route->setUriPattern($uriPattern);
-            if (isset($routeConfiguration['defaults'])) {
-                $route->setDefaults($routeConfiguration['defaults']);
-            }
-            if (isset($routeConfiguration['routeParts'])) {
-                $route->setRoutePartsConfiguration($routeConfiguration['routeParts']);
-            }
-            if (isset($routeConfiguration['toLowerCase'])) {
-                $route->setLowerCase($routeConfiguration['toLowerCase']);
-            }
-            if (isset($routeConfiguration['appendExceedingArguments'])) {
-                $route->setAppendExceedingArguments($routeConfiguration['appendExceedingArguments']);
-            }
-            if (isset($routeConfiguration['httpMethods'])) {
-                if (isset($routesWithHttpMethodConstraints[$uriPattern]) && $routesWithHttpMethodConstraints[$uriPattern] === false) {
-                    throw new InvalidRouteSetupException(sprintf('There are multiple routes with the uriPattern "%s" and "httpMethods" option set. Please specify accepted HTTP methods for all of these, or adjust the uriPattern', $uriPattern), 1365678427);
-                }
-                $routesWithHttpMethodConstraints[$uriPattern] = true;
-                $route->setHttpMethods($routeConfiguration['httpMethods']);
-            } else {
-                if (isset($routesWithHttpMethodConstraints[$uriPattern]) && $routesWithHttpMethodConstraints[$uriPattern] === true) {
-                    throw new InvalidRouteSetupException(sprintf('There are multiple routes with the uriPattern "%s" and "httpMethods" option set. Please specify accepted HTTP methods for all of these, or adjust the uriPattern', $uriPattern), 1365678432);
-                }
-                $routesWithHttpMethodConstraints[$uriPattern] = false;
-            }
-            $this->routes[] = $route;
-        }
-        $this->routesCreated = true;
-    }
-
-    /**
-     * Checks if a routes configuration was set and otherwise loads the configuration from the configuration manager.
-     *
-     * @return void
-     */
-    protected function initializeRoutesConfiguration()
-    {
-        if ($this->routesConfiguration === null) {
-            $this->routesConfiguration = $this->configurationManager->getConfiguration(ConfigurationManager::CONFIGURATION_TYPE_ROUTES);
-        }
     }
 }
